+++ test/integration/recommend-page.js
@@ -11,6 +11,7 @@ chai.use(chaiHttp);
 
 const recommendRoute = `${constants.SITE_ROOT}/recommend`;
 const chooseRoute = `${constants.SITE_ROOT}/choose`;
+const locationSearchRoute = `${constants.SITE_ROOT}/location`;
 
 describe('Recommend page', () => {
   it('page title should be \'We recommend that you\' if symptoms question is answered yes', async () => {
@@ -35,17 +36,35 @@ describe('Recommend page', () => {
 
   describe('return to Choices services', () => {
     it('the breadcrumb should have a link back to the Choices \'Services near you\'', async () => {
-      const res = await chai.request(server).get(recommendRoute);
+      const res = await chai.request(server)
+        .get(recommendRoute)
+        .query({ symptoms: 'yes' });
 
+      iExpect.htmlWith200Status(res);
       const $ = cheerio.load(res.text);
 
       expect($($('div.breadcrumb a')[1]).attr('href'))
         .to.equal('https://www.nhs.uk/service-search');
     });
 
-    it('the page should have a link back to the Choices service search', async () => {
-      const res = await chai.request(server).get(recommendRoute);
+    it('the page should have a link to the Location search page', async () => {
+      const res = await chai.request(server)
+        .get(recommendRoute)
+        .query({ symptoms: 'yes' });
 
+      iExpect.htmlWith200Status(res);
+      const $ = cheerio.load(res.text);
+
+      expect($('.button').attr('href'))
+        .to.equal(locationSearchRoute);
+    });
+
+    it('the page should have a link back to the Choices service search', async () => {
+      const res = await chai.request(server)
+        .get(recommendRoute)
+        .query({ symptoms: 'yes' });
+
+      iExpect.htmlWith200Status(res);
       const $ = cheerio.load(res.text);
 
       expect($('.back-to-choices').attr('href'))