+++ app/middleware/selectionValidatorAge.js
@@ -5,6 +5,8 @@ function ageSelection(req, res, next) {
     routeHelper.renderAgePageWithError(req, res);
   } else if (res.locals.age === '1') {
     routeHelper.renderSeeExpertUnder16Page(req, res);
+  } else if (res.locals.age === '2') {
+    routeHelper.renderChooseTestUnder25Page(req, res);
   } else {
     next();
   }