+++ app/public/js/init.js
@@ -1,2 +1,3 @@
 // Require all modules needed for the application
 require('./cookieMessage');
+require('./radioSelect');